feat: Add executor submit_firmware
