+++ embark/uploader/settings.py
@@ -1,0 +1,22 @@
+from django.conf import settings
+
+
+def workers_enabled():
+    # Todo: Replace with actual check whether workers are enabled
+    return False
+
+
+def get_emba_root():
+    """
+    Gets EMBA root considering if workers are enabled
+    :returns: path to emba folder
+    """
+    return settings.WORKER_EMBA_ROOT if workers_enabled() else settings.EMBA_ROOT
+
+
+def get_emba_base_cmd():
+    """
+    Constructs EMBA base command
+    :returns: EMBA base command
+    """
+    return f"sudo DISABLE_STATUS_BAR=1 DISABLE_NOTIFICATIONS=1 HTML=1 FORMAT_LOG=1 {get_emba_root()}/emba"
