+++ embark/uploader/views.py
@@ -17,7 +17,7 @@ from rest_framework.response import Response
 from rest_framework.parsers import MultiPartParser
 from rest_framework import serializers
 
-from uploader.boundedexecutor import BoundedExecutor
+from uploader.executor import submit_firmware
 from uploader.forms import DeviceForm, FirmwareAnalysisForm, DeleteFirmwareForm, LabelForm, VendorForm
 from uploader.models import FirmwareFile
 from uploader.serializers import FirmwareAnalysisSerializer
@@ -145,7 +145,7 @@ def start_analysis_serialized(data):
     logger.debug("new_analysis %s has label: %s", new_analysis, new_analysis.label)
 
     # inject into bounded Executor
-    if not BoundedExecutor.submit_firmware(firmware_analysis=new_analysis, firmware_file=new_firmware_file):
+    if not submit_firmware(firmware_analysis=new_analysis, firmware_file=new_firmware_file):
         raise BufferFullException
 
     return new_analysis.id
@@ -254,7 +254,7 @@ def start_analysis(request):
             new_analysis.save()
             logger.debug("new_analysis %s has label: %s", new_analysis, new_analysis.label)
             # inject into bounded Executor
-            if BoundedExecutor.submit_firmware(firmware_analysis=new_analysis, firmware_file=new_firmware_file):
+            if submit_firmware(firmware_analysis=new_analysis, firmware_file=new_firmware_file):
                 return redirect('embark-dashboard-service')
             logger.error("Server Queue full, or other boundenexec error")
             return HttpResponseServerError("Queue full")
